requirements.txt
backports.shutil-get-terminal-size==1.0.0
cffi==1.7.0
cryptography==1.5
decorator==4.0.10
Django==1.10.5
django-appconf==1.0.2
django-extensions==1.7.3
django-ipware==1.1.5
django-registration==2.1.2
Django-Select2==5.8.9
enum34==1.1.6
idna==2.1
ipaddress==1.0.16
ipython==5.1.0
ipython-genutils==0.1.0
MySQL-python==1.2.5
ndg-httpsclient==0.4.2
nltk==3.2.1
numpy==1.11.1
pandas==0.18.1
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.4.1
prompt-toolkit==1.0.7
ptyprocess==0.5.1
pyasn1==0.1.9
pycparser==2.14
Pygments==2.1.3
pyocr==0.4.2
pyOpenSSL==16.1.0
pytesseract==0.1.6
python-dateutil==2.5.3
python-memcached==1.58
pytz==2016.6.1
requests==2.11.1
short-url==1.2.2
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.2
Wand==0.4.3
wcwidth==0.1.7
yolk==0.4.3
